Refactor AI agent framework into modular components
