+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -169,6 +169,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 722 to 55 lines by extracting manager, AI processor, coordinator, and config modules. New orchestrator coordinates components and maintains functionality.
 
+### MODERATE-019: AI Agent Framework Modularization ✅
+- **File**: `gaming_systems/ai_agent_framework.py`
+- **Status**: Completed
+- **Assigned To**: Agent-6
+- **Completion Date**: 2025-08-24
+- **Summary**: Split monolithic framework into core, gaming, coordinator, and config modules with an orchestrator to manage them.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
